git-svn-id: https://svn.apache.org/repos/asf/perl/modperl/branches/clone-skip-unstable@209866 13f79535-47bb-0310-9956-ffa450edef68

+++ xs/ModPerl/Util/ModPerl__Util.h
@@ -18,7 +18,7 @@
     Perl_newSVpvf(aTHX_ "0x%lx", (unsigned long)aTHX)
 #else
 #define mpxs_ModPerl__Util_current_perl_id() \
-    Perl_newSVpvf(aTHX_ "0x%lx", 0)
+    Perl_newSVpvf(aTHX_ "0x%lx", (unsigned long)0)
 #endif
 
 static MP_INLINE void mpxs_ModPerl__Util_untaint(pTHX_ I32 items,
